Implementation of PseudoSensor (#33)
The PR introduced a pseudo sensor for the position of goals and obstacles. The exact information about the location of obstacles and goals can be accessed through the observation. If the sensor is added, obstacles and goals are accessible through the keys `ObstaclePosition` and `GoalPosition`. 
The current implementation only supports sensors for the pointMass, where the workspaces equals the configuration space.